restructure seed operator tests
